MR-199: Navbar in header
