+++ src/Components/Layout/Header/NavBar/index.jsx
@@ -1,12 +1,33 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styles from './navbar.module.css';
+import Button from 'Components/Shared/Button';
+import { useHistory } from 'react-router-dom/cjs/react-router-dom.min';
 
 const NavBar = ({ routes = [] }) => {
+  const role = sessionStorage.getItem('role');
+  const history = useHistory();
+
+  const privateRoutes = routes.filter((route) => route.name !== 'Home');
+
   return (
     <nav className={styles.navbar}>
+      <Link to={routes[0].path}>
+        <div data-testid="logo-container" className={styles.logoContainer}>
+          <img
+            src={`${process.env.PUBLIC_URL}/assets/images/logo.png`}
+            alt="Flexing on fire logo"
+            className={styles.logo}
+          />
+          <img
+            src={`${process.env.PUBLIC_URL}/assets/images/logo2.png`}
+            alt="MegaRocket words logo"
+            className={styles.logo2}
+          />
+        </div>
+      </Link>
       <ul className={styles.routes} data-testid="routes-list">
-        {routes.map((route) => (
+        {privateRoutes.map((route) => (
           <li key={route.name}>
             <Link to={route.path} className={styles.a}>
               {route.name}
@@ -14,6 +35,22 @@ const NavBar = ({ routes = [] }) => {
           </li>
         ))}
       </ul>
+      <div className={styles.buttonContainer} data-testid="home-buttons-container">
+        {!role && (
+          <>
+            <Button
+              text="Sign Up"
+              classNameButton="submitButton"
+              action={() => history.push('/auth/signup')}
+            />
+            <Button
+              text="Log In"
+              classNameButton="submitButton"
+              action={() => history.push('/auth/login')}
+            />
+          </>
+        )}
+      </div>
     </nav>
   );
 };
