+++ src/Components/Layout/Header/NavBar/navbar.module.css
@@ -1,26 +1,38 @@
 .navbar {
   display: flex;
+  justify-content: space-between;
   align-items: center;
-  justify-content: center;
-  padding: 0 5%;
-  min-height: 70px;
-  background-color: #D9D9D9;
+  width: 100%;
+}
+
+.logoContainer{
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+}
+.logo {
+  width: 80px;
+}
+
+.logo2 {
+  width: 180px;
+  height: 38px;
+}
+
+.buttonContainer{
+  display: flex;
 }
 
 .routes {
   display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  width: 90%;
-  padding: 0;
   list-style-type: none;
 }
 
 .a {
-  margin: 0 10px;
-  font-size: 1rem;
+  margin-left: 25px;
+  font-size: 1.2rem;
   font-weight: 600;
   text-decoration: none;
   text-transform: capitalize;
-  color: black;
+  color: #fff;
 }
