+++ src/Components/Layout/Header/header.module.css
@@ -2,29 +2,14 @@
   position: fixed;
   width: 100%;
   display: flex;
-  justify-content: space-between;
+  align-items: center;
   background-color: #373867;
 }
 
 .container2{
   display: flex;
   align-items: center;
-  justify-content: space-around;
   padding-right: 2%;
-}
-
-.logoContainer{
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-.logo {
-  width: 80px;
-}
-
-.logo2 {
-  width: 180px;
-  height: 38px;
 }
 
 .socialIcon {
@@ -32,14 +17,6 @@
 }
 .logOutIcon{
   margin-left: 5px;
-}
-.navbar {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0 5%;
-  min-height: 70px;
-  background-color: #D9D9D9;
 }
 
 .appName {
@@ -50,8 +27,6 @@
 .rutes {
   display: flex;
   flex-wrap: wrap;
-  justify-content: space-evenly;
-  width: 90%;
   padding: 0;
   list-style-type: none;
 }
