+++ src/Routes/auth/index.jsx
@@ -9,14 +9,6 @@ const routes = [
   {
     name: 'Home',
     path: '/'
-  },
-  {
-    name: 'Sign Up',
-    path: '/auth/signup'
-  },
-  {
-    name: 'Log In',
-    path: '/auth/login'
   }
 ];
